// src/enums/earning.enum.ts
export enum PaymentType {
  ADMISSION_FEE = "Admission Fee",
  FORM_FEE = "Form Fee",
  MONTHLY_FEE = "Monthly Fee",
}

export enum Months {
  JANUARY = "January",
  FEBRUARY = "February",
  MARCH = "March",
  APRIL = "April",
  MAY = "May",
  JUNE = "June",
  JULY = "July",
  AUGUST = "August",
  SEPTEMBER = "September",
  OCTOBER = "October",
  NOVEMBER = "November",
  DECEMBER = "December",
}